BX-1670 websockets transaction feed (#883)
Also addresses BX-1672 and BX-1676.

Primary changes:

 - Implement websocket server that handles JSON-RPC calls
 - Implement transaction pub/sub interface for websocket connections

This particular PR doesn't really contain the above changes, but includes relevant refactoring changes:

 - Refactored all RPC logic. RPC handlers are now general JSON-RPC handlers and not HTTP specific; websocket and HTTP RPC handlers will call the same underlying classes.
 - Refactored the usage of opts (CommonOpts, GatewayOpts, etc.), and made safe type annotations of this attribute mandatory.
 - Added full-stack integration tests for HTTP RPC interfaces.

+++ src/bxcommon/services/transaction_service.py
@@ -1174,7 +1174,7 @@ class TransactionService:
         """
         if self.node.opts.transaction_pool_memory_limit is not None:
             # convert MB to bytes
-            return self.node.opts.transaction_pool_memory_limit * 1024 * 1024
+            return int(self.node.opts.transaction_pool_memory_limit * 1024 * 1024)
 
         for blockchain_network in self.node.opts.blockchain_networks:
             if blockchain_network.network_num == self.network_num:
